Move float8_aten_api to float8_ops
