+++ torchao/float8/float8_ops.py
@@ -6,12 +6,13 @@
 from typing import Any, Dict, Tuple
 
 import torch
+from torch.library import Library
+
+from torch.utils._pytree import tree_map
 
 from torchao.float8.float8_python_api import addmm_float8_unwrapped
 from torchao.float8.float8_tensor import choose_scaled_mm_config, Float8Tensor
 from torchao.float8.float8_utils import is_row_major, pad_tensor_for_matmul
-
-from torch.utils._pytree import tree_map
 
 aten = torch.ops.aten
 c10d_functional = torch.ops.c10d_functional
@@ -155,6 +156,7 @@ def float8_view(aten_op, args, kwargs=None):
 def float8_split(aten_op, args, kwargs=None):
     new_data_tensors = aten_op(args[0]._data, *args[1:], **kwargs)
     _assert_tensorwise_scale(aten_op, args[0]._scale)
+
     def make_float8(data):
         return Float8Tensor(
             data,
@@ -481,3 +483,36 @@ def copy_fp8(aten_op, args, kwargs=None):
         )
     else:
         raise RuntimeError("Unsupported semantics for copy_ in Float8Tensor")
+
+
+def mm_float8_emulated(
+    m1,  # input 1 data
+    s1,  # input 1 scale
+    m2,  # input 2 data
+    s2,  # input 2 scale
+    dtype3,  # output dtype
+):
+    # naive implementation: dq -> op -> q
+    m1_fp32 = m1.float() / s1
+    m2_fp32 = m2.float() / s2
+    m3_fp32 = torch.mm(m1_fp32, m2_fp32)
+
+    return m3_fp32.to(dtype3)
+
+
+# ATen op placeholders
+# Register the aten level functions we need.
+# These are mostly placeholder and might need to be implemented in c++ as needed
+lib = Library("aten", "FRAGMENT")
+
+lib.define(
+    "mm_float8_emulated(Tensor m1, Tensor s1, Tensor m2, Tensor s2, ScalarType dtype3) -> Tensor"
+)
+lib.impl("mm_float8_emulated", mm_float8_emulated, "CPU")
+lib.impl("mm_float8_emulated", mm_float8_emulated, "CUDA")
+
+
+@torch.library.impl(lib, "mm_float8_emulated", "Meta")
+def _mm_float8_emulated_meta(m1, s1, m2, s2, dtype3):
+    out = torch.mm(m1.float(), m2.float()).to(dtype3)
+    return out
